+++ src/test/java/org/greencodeinitiative/creedengo/python/checks/AvoidWildcardImportsCheckTest.java
@@ -15,45 +15,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.greencodeinitiative.creedengo.python.integration.tests.profile;
+package org.greencodeinitiative.creedengo.python.checks;
 
-import java.util.List;
+import org.junit.jupiter.api.Test;
+import org.sonar.python.checks.utils.PythonCheckVerifier;
 
-public class ProfileMetadata {
-	private String name;
-	private String language;
-	private List<String> ruleKeys;
+public class AvoidWildcardImportsCheckTest {
 
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getLanguage() {
-		return language;
-	}
-
-	public void setLanguage(String language) {
-		this.language = language;
-	}
-
-	public List<String> getRuleKeys() {
-		return ruleKeys;
-	}
-
-	public void setRuleKeys(List<String> ruleKeys) {
-		this.ruleKeys = ruleKeys;
-	}
-
-	@Override
-	public String toString() {
-		return "ProfileMetadata{" +
-				"name='" + name + '\'' +
-				", language='" + language + '\'' +
-				", ruleKeys=" + ruleKeys +
-				'}';
-	}
+    @Test
+    public void test() {
+        PythonCheckVerifier.verify("src/test/resources/checks/avoidWildcardImports.py", new AvoidWildcardImportsCheck());
+    }
 }
+